requirements.txt
# Core dependencies
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
pandas>=1.3.0
pyyaml>=5.4.0

# Apache DataSketches
datasketches>=3.0.0

# Visualization (optional)
matplotlib>=3.4.0
seaborn>=0.11.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
